preserve wide ERAY scheduler periods

+++ include/hw/net/tricore_eray.h
@@ -40,7 +40,8 @@ typedef struct TriCoreERAYState {
     uint32_t last_rx_id, last_rx_cycle;
     uint8_t last_rx_channel;
     uint32_t sched_cfg;
-    uint32_t sched_period_ns;
+    /* GTU products exceed 32 bits on maximum public settings. */
+    uint64_t sched_period_ns;
     uint32_t slot_counter, minislot_counter;
     uint32_t tx_frame_id, tx_due_cycle, tx_due_slot, tx_payload_len;
     uint32_t tx_header_flags;
